+++ src/configuration.ts
@@ -88,6 +88,13 @@ export class Configuration {
   }
 
   /**
+   * Whether Format Pending Changes should include staged changes.
+   */
+  public getFormatPendingChangesIncludeStaged(scope?: vscode.Uri): boolean {
+    return this.getValue(scope, 'formatPendingChanges.includeStaged', true);
+  }
+
+  /**
    * Whether to run with `bundle exec`.
    */
   public getPreferBundler(formatter: FormatterName, scope?: vscode.Uri): boolean {